scheduler/requirements.txt
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0